+++ src/ticks/syms.c
@@ -105,7 +105,7 @@ void read_symbol_file(char *filename)
                      strncasecmp(argv[0], "__data", 6) == 0 ||
                      strncasecmp(argv[0], "__rodata", 8) == 0 ) {
                     // It's something to do with a section 
-                    int len = strlen(argv[0]);
+                    int len = (int)strlen(argv[0]);
 
                     // Rely on z80asm writing them out, __head, __tail, __size
                     // Once we detect __size, we have all the informaiton we need
@@ -368,7 +368,7 @@ const char *find_symbol(int addr, symboltype preferred_type)
 char **parse_words(char *line, int *argc)
 {
     int                 i = 0, j = 0 , n = 0;
-    int                 len = strlen(line);
+    int                 len = (int)strlen(line);
     int                 in_single_quotes = 0, in_double_quotes = 0;
     char              **args;
